Refactor: Extract integrations page into modular components
Reduced the monolithic integrations page from 6,504 lines to 3,860 lines (40.6% reduction) by extracting reusable components, handlers, and utilities.

**Changes:**
- Extracted 22 new modules organized into handlers/, components/, and dialogs/
- Created types.ts and utils.ts for shared code
- Improved code maintainability and reusability

**Components extracted:**
- GitHubIntegrationCard, SlackIntegrationCard
- GitHubConnectedCard, SlackConnectedCard
- RootlyIntegrationForm, PagerDutyIntegrationForm
- DeleteIntegrationDialog, GitHubDisconnectDialog, SlackDisconnectDialog
- NewMappingDialog, OrganizationManagementDialog

**Handlers extracted:**
- github-handlers.ts, slack-handlers.ts
- integration-handlers.ts, mapping-handlers.ts
- team-handlers.ts, organization-handlers.ts, ai-handlers.ts
